I-READ-BackendAPI/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
async-timeout==5.0.1
attrs==24.2.0
azure-cognitiveservices-speech==1.41.1
black==24.10.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
crispy-tailwind==1.0.3
deprecation==2.1.0
dj-database-url==2.3.0
Django==5.1.2
django-autoslug==1.9.9
django-cors-headers==4.6.0
django-crispy-forms==2.3
django-environ==0.11.2
django-filter==24.3
django-ninja==1.3.0
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
et_xmlfile==2.0.0
frozenlist==1.5.0
gotrue==2.10.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.6
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
keyboard==0.13.5
Markdown==3.7
MouseInfo==0.1.3
multidict==6.1.0
mypy-extensions==1.0.0
numpy==2.1.2
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
pathspec==0.12.1
Pillow==9.4.0
platformdirs==4.3.6
postgrest==0.17.2
propcache==0.2.0
psycopg==3.2.3
psycopg-binary==3.2.3
psycopg2==2.9.10
PyAutoGUI==0.9.53
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
PyGetWindow==0.0.9
PyJWT==2.10.1
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytweening==1.0.4
pytz==2024.2
realtime==2.0.6
redis==5.2.0
requests==2.32.3
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
storage3==0.8.2
supabase==2.9.1
supafunc==0.6.2
typing_extensions==4.12.2
tzdata==2024.2
upstash-redis==1.2.0
urllib3==2.2.3
websockets==13.1
yarl==1.17.1
fuzzywuzzy==0.18.0
gunicorn==21.2.0
whitenoise==6.6.0
